move TimezoneComponent to Locations namespace
