+++ app/views/developers/_details.html.erb
@@ -26,7 +26,7 @@
 
   <%= render RenderableComponent.new("space-y-3 pt-4") do %>
     <%= render Locations::LocationComponent.new(developer.location) %>
-    <%= render TimeZoneComponent.new(developer.location) %>
+    <%= render Locations::TimeZoneComponent.new(developer.location) %>
   <% end %>
 
   <%= render RenderableComponent.new("space-y-3 pt-4") do %>
